Switch database from Postgres to DynamoDB
The dataset is growing so let's try using DynamoDB. For now, we're
using both boto (for DynamoDB operations) and botocore (for querying
spot price data). Eventually, we'll move everything to botocore and
possibly use the pynamodb interface. Also:

* Start collecting data for all EC2 "products":

  * Linux/UNIX
  * Linux/UNIX (Amazon VPC)
  * SUSE Linux
  * SUSE Linux (Amazon VPC)
  * Windows
  * Windows (Amazon VPC)

  Previously, we only collected data for Linux/UNIX.

* Handle next_token properly to pull spot price data that is more
  than one page.

* Save the "end_time" timestamp of each run, so we know where to
  start from next time.

* Update to latest bootstrap, jquery, and nvd3.

* Tweak timestamp on graph X axis to include HH:mm.

Performance-wise, the app still feels about the same as when loading
data from Postgres. Possibly it's a little bit slower. This can be
tuned in a future change. We might even use memcache to reduce the
number of DynamoDB operations.